Fix failing tests (#2726)
